feat: add professional README for core-onsocial contract
- Add clean, concise README with API overview and quick start
- Document unified execute API, actions, auth modes, and views
- Include build and test commands using existing Makefile targets
- Remove verbose documentation in favor of developer-friendly format

+++ contracts/core-onsocial/src/tests/unit/event_emission_test.rs
@@ -134,7 +134,7 @@ mod event_emission_tests {
                         let partition = data.partition_id.unwrap();
 
                         assert!(
-                            partition < NUM_PARTITIONS as u16,
+                            partition < NUM_PARTITIONS,
                             "Partition ID should be < {}",
                             NUM_PARTITIONS
                         );
